+++ precompiles/sha256f/src/lib.rs
@@ -6,8 +6,6 @@ mod sha256f_input;
 mod sha256f_instance;
 mod sha256f_manager;
 mod sha256f_planner;
-mod sha256f_table;
-mod sha256f_types;
 
 pub use sha256f::*;
 pub use sha256f_bus_device::*;
@@ -17,5 +15,3 @@ pub use sha256f_input::*;
 pub use sha256f_instance::*;
 pub use sha256f_manager::*;
 pub use sha256f_planner::*;
-pub use sha256f_table::*;
-pub use sha256f_types::*;
