+++ precompiles/sha256f/src/sha256f.rs
@@ -1,89 +1,45 @@
 use core::panic;
-use std::{fs, path::PathBuf, sync::Arc};
+use std::sync::Arc;
 
 use fields::PrimeField64;
-use generic_array::{typenum::U64, GenericArray};
-use sha2::compress256;
-
-use proofman_common::{AirInstance, FromTrace, SetupCtx};
-use proofman_util::{timer_start_trace, timer_stop_and_log_trace};
-use zisk_core::{convert_u64_to_generic_array_bytes, convert_u64_to_u32};
-use zisk_pil::{Sha256fFixed, Sha256fTrace, Sha256fTraceRow};
-
-use crate::Sha256fInput;
-
-use super::{sha256f_constants::*, InputType, Script, Sha256fTableGateOp, Sha256fTableSM};
-
 use rayon::prelude::*;
+
+use pil_std_lib::Std;
+use proofman_common::{AirInstance, FromTrace};
+use proofman_util::{timer_start_trace, timer_stop_and_log_trace};
+use zisk_pil::{Sha256fTrace, Sha256fTraceRow};
+
+use super::{sha256f_constants::*, Sha256fInput};
 
 /// The `Sha256fSM` struct encapsulates the logic of the Sha256f State Machine.
 pub struct Sha256fSM<F: PrimeField64> {
-    /// Reference to the Sha256f Table State Machine.
-    sha256f_table_sm: Arc<Sha256fTableSM>,
-
-    /// Script for the Sha256f's circuit representation
-    script: Arc<Script>,
-
-    /// Size of a circuit in the trace. It corresponds to the number of gates in the circuit.
-    circuit_size: usize,
-
-    /// Number of available circuits in the trace.
-    num_available_circuits: usize,
+    /// Reference to the PIL2 standard library.
+    pub std: Arc<Std<F>>,
 
     /// Number of available sha256fs in the trace.
     pub num_available_sha256fs: usize,
 
-    sha256f_fixed: Sha256fFixed<F>,
+    num_non_usable_rows: usize,
+
+    /// Range checks ID's
+    a_range_id: usize,
+    e_range_id: usize,
 }
 
 impl<F: PrimeField64> Sha256fSM<F> {
     /// Creates a new Sha256f State Machine instance.
     ///
-    /// # Arguments
-    /// * `sha256f_table_sm` - An `Arc`-wrapped reference to the Sha256f Table State Machine.
-    ///
     /// # Returns
     /// A new `Sha256fSM` instance.
-    pub fn new(
-        sctx: Arc<SetupCtx<F>>,
-        sha256f_table_sm: Arc<Sha256fTableSM>,
-        script_path: PathBuf,
-    ) -> Arc<Self> {
-        let script = fs::read_to_string(script_path).expect("Failed to read sha256f_script.json");
-        let script: Script =
-            serde_json::from_str(&script).expect("Failed to parse sha256f_script.json");
-
-        // Get the circuit size
-        let circuit_size = script.total;
-        let circuit_ops_count = &script.sums;
-
-        // Check that the script is valid
-        assert!(
-            circuit_ops_count.xor
-                + circuit_ops_count.ch
-                + circuit_ops_count.maj
-                + circuit_ops_count.add
-                == circuit_size + 1
-        );
-        assert!(script.program.len() == circuit_size);
-
+    pub fn new(std: Arc<Std<F>>) -> Arc<Self> {
         // Compute some useful values
-        let num_available_circuits = (Sha256fTrace::<usize>::NUM_ROWS - 1) / circuit_size;
-        let num_available_sha256fs = NUM_SHA256F_PER_CIRCUIT * num_available_circuits;
+        let num_available_sha256fs = Sha256fTrace::<usize>::NUM_ROWS / CLOCKS - 1;
+        let num_non_usable_rows = Sha256fTrace::<usize>::NUM_ROWS % CLOCKS;
 
-        let airgroup_id = Sha256fTrace::<usize>::AIRGROUP_ID;
-        let air_id = Sha256fTrace::<usize>::AIR_ID;
-        let fixed_pols = sctx.get_fixed(airgroup_id, air_id);
-        let sha256f_fixed = Sha256fFixed::from_vec(fixed_pols);
+        let a_range_id = std.get_range(0, (1 << 3) - 1, None);
+        let e_range_id = std.get_range(0, (1 << 3) - 1, None);
 
-        Arc::new(Self {
-            sha256f_table_sm,
-            script: Arc::new(script),
-            circuit_size,
-            num_available_circuits,
-            num_available_sha256fs,
-            sha256f_fixed,
-        })
+        Arc::new(Self { std, num_available_sha256fs, num_non_usable_rows, a_range_id, e_range_id })
     }
 
     /// Processes a slice of operation data, updating the trace and multiplicities.
@@ -94,428 +50,286 @@ impl<F: PrimeField64> Sha256fSM<F> {
     /// * `input` - The operation data to process.
     /// * `multiplicity` - A mutable slice to update with multiplicities for the operation.
     #[inline(always)]
-    pub fn process_trace<'a, I>(
-        &self,
-        trace: &mut Sha256fTrace<F>,
-        num_rows_constants: usize,
-        inputs: I,
-        num_inputs: usize,
-    ) where
-        I: IntoIterator<Item = &'a Sha256fInput>,
-    {
-        let mut inputs_bits: Vec<[u64; INPUT_DATA_SIZE_BITS]> =
-            vec![[0u64; INPUT_DATA_SIZE_BITS]; self.num_available_circuits];
+    pub fn process_input(&self, input: &Sha256fInput, trace: &mut [Sha256fTraceRow<F>]) {
+        let step_main = input.step_main;
+        let addr_main = input.addr_main;
+        let state_addr = input.state_addr;
+        let input_addr = input.input_addr;
+        let state = &input.state;
+        let input = &input.input;
 
-        // Process the inputs
-        let initial_offset = num_rows_constants;
-        let input_offset = INPUT_SIZE;
-        let output_offset = OUTPUT_SIZE;
-        let mut circuit = 0;
-        for (i, input) in inputs.into_iter().enumerate() {
-            // Get the basic data from the input
-            let step_received = input.step_main;
-            let addr_received = input.addr_main;
-            let state_addr_received = input.state_addr;
-            let input_addr_received = input.input_addr;
-            let state_received = &input.state;
-            let input_received = &input.input;
+        // Fill the step_addr
+        trace[0].step_addr = F::from_u64(step_main); // STEP_MAIN
+        trace[1].step_addr = F::from_u32(addr_main); // ADDR_OP
+        trace[2].step_addr = F::from_u32(state_addr); // ADDR_STATE
+        trace[3].step_addr = F::from_u32(input_addr); // ADDR_INPUT
+        trace[4].step_addr = F::from_u32(state_addr); // ADDR_IND_0
+        trace[5].step_addr = F::from_u32(input_addr); // ADDR_IND_1
 
-            // Convert the input data to 32-bit words
-            let mut state_u32: [u32; 8] = convert_u64_to_u32(state_received).try_into().unwrap();
-            let input_u32: [u32; 16] = convert_u64_to_u32(input_received).try_into().unwrap();
+        // Activate the clk_0 selector
+        trace[0].in_use_clk_0 = F::ONE;
 
-            // And collect them
-            let mut sha256f_input_data = [0u32; INPUT_DATA_SIZE_U32];
-            sha256f_input_data[..8].copy_from_slice(&state_u32);
-            sha256f_input_data[8..].copy_from_slice(&input_u32);
-
-            circuit = i / NUM_SHA256F_PER_CIRCUIT;
-            let circuit_pos = i % NUM_SHA256F_PER_CIRCUIT;
-            let circuit_offset = circuit * self.circuit_size;
-
-            // Update the multiplicity for the input
-            let initial_pos = initial_offset + circuit_offset + circuit_pos;
-            trace[initial_pos].in_use_clk_0 = F::ONE; // The pair (step_received, addr_received) is unique each time
-
-            // Update the step_addr and in_use as expected
-            let mut offset = initial_pos;
-            for j in 0..IN_DATA_BLOCKS {
-                let rb_offset = j * RB_SIZE;
-                let pos = offset + rb_offset;
-
-                trace[pos].in_use = F::ONE;
-
-                trace[pos].step_addr = match j {
-                    0 => F::from_u64(step_received),       // STEP_MAIN
-                    1 => F::from_u32(addr_received),       // ADDR_OP
-                    2 => F::from_u32(state_addr_received), // ADDR_STATE
-                    3 => F::from_u32(input_addr_received), // ADDR_INPUT
-                    4 => F::from_u32(state_addr_received), // ADDR_IND_0
-                    5 => F::from_u32(input_addr_received), // ADDR_IND_1
-                    _ => F::ZERO,
-                };
-            }
-
-            // Process the input data
-            sha256f_input_data.iter().enumerate().for_each(|(j, &value)| {
-                let block_offset = j * BLOCK_SIZE;
-                let pos = offset + block_offset;
-
-                // Process the 32-bit chunk
-                for k in 0..32 {
-                    let bit = ((value >> k) & 1) as u64;
-
-                    // Divide the value in bits:
-                    //    (circuit i) [0b1011,  0b0011,  0b1000,  0b0010]
-                    //    (circuit i) [1,0,1,1, 0,0,1,1, 1,0,0,0, 0,0,1,0]
-                    let bit_pos = k + 32 * j;
-                    let old_value = inputs_bits[circuit][bit_pos];
-                    inputs_bits[circuit][bit_pos] = (bit << circuit_pos) | old_value;
-
-                    // We update bit[i] and val[i]
-                    let bit_pos = k % BITS_IN_PARALLEL;
-                    let bit_offset = (k - bit_pos) * NUM_SHA256F_PER_CIRCUIT / BITS_IN_PARALLEL;
-                    update_bit_val(
-                        trace,
-                        pos + bit_offset,
-                        bit,
-                        circuit_pos,
-                        bit_pos,
-                        circuit_pos == 0 && (pos + bit_offset > 1),
-                    );
-                }
-            });
-
-            // Activate the in_use for the output data
-            offset += input_offset;
-            for j in 0..OUT_BLOCKS {
-                let rb_offset = j * RB_SIZE;
-                let pos = offset + rb_offset;
-
-                trace[pos].in_use = F::ONE;
-            }
-
-            // Obtain the sha256f output
-            let block: GenericArray<u8, U64> = convert_u64_to_generic_array_bytes(input_received);
-            compress256(&mut state_u32, &[block]);
-
-            // Process the output
-            state_u32.iter().enumerate().for_each(|(j, &value)| {
-                let block_offset = j * BLOCK_SIZE;
-                let pos = offset + block_offset;
-
-                // Process the 32-bit chunk
-                for k in 0..32 {
-                    let bit = ((value >> k) & 1) as u64;
-
-                    // We update bit[i] and val[i]
-                    let bit_pos = k % BITS_IN_PARALLEL;
-                    let bit_offset = (k - bit_pos) * NUM_SHA256F_PER_CIRCUIT / BITS_IN_PARALLEL;
-                    update_bit_val(
-                        trace,
-                        pos + bit_offset,
-                        bit,
-                        circuit_pos,
-                        bit_pos,
-                        circuit_pos == 0,
-                    );
-                }
-            });
-
-            // Finally, activate the in_use for the two indirections
-            offset += output_offset;
-            for j in 0..2 {
-                let rb_offset = j * RB_SIZE;
-                let pos = offset + rb_offset;
-
-                trace[pos].in_use = F::ONE;
-            }
+        // Activate the in_use selector
+        for r in trace.iter_mut().take(18) {
+            r.in_use = F::ONE;
         }
 
-        // It the number of inputs is less than the available sha256fs, we need to fill the remaining inputs
-        if num_inputs < self.num_available_sha256fs {
-            // Compute the hash of zero
-            let mut zero_state = [0u32; 8];
-            let block_zeros: GenericArray<u8, U64> = GenericArray::default();
-            compress256(&mut zero_state, &[block_zeros]);
-            // hash_of_0: [0x7ca51614, 0x425c3ba8, 0xce54dd2f, 0xc2020ae7, 0xb6e574d1, 0x98136d0f, 0xae7e26cc, 0xbf0be7a6]
+        // Compute the load state stage
+        let mut offset = 0;
+        let mut prev_state = [0u32; 8];
+        for i in 0..CLOCKS_LOAD_STATE {
+            let word = state[i];
+            let word_high = (word >> 32) as u32;
+            let word_low = (word & 0xFFFF_FFFF) as u32;
 
-            // If the number of inputs is not a multiple of NUM_SHA256F_PER_CIRCUIT,
-            // we fill the last processed circuit
-            let rem_inputs = num_inputs % NUM_SHA256F_PER_CIRCUIT;
-            if rem_inputs != 0 {
-                let circuit_offset = circuit * self.circuit_size;
+            // Store the state as u32 for further processing
+            prev_state[2 * i] = word_high;
+            prev_state[2 * i + 1] = word_low;
 
-                // Since no more bits are being introduced as input, we let 0 be the
-                // new bits and therefore we repeat the last values
-                let mut offset = initial_offset + circuit_offset;
-                for j in 0..INPUT_DATA_SIZE_BITS / BITS_IN_PARALLEL {
-                    let block_offset = j * NUM_SHA256F_PER_CIRCUIT;
-                    let block = offset + block_offset;
-                    for k in rem_inputs..NUM_SHA256F_PER_CIRCUIT {
-                        let pos = block + k;
-                        for l in 0..BITS_IN_PARALLEL {
-                            trace[pos + 1].val[l] = trace[pos].val[l];
-                        }
-                    }
-                }
+            let mut row = if i == 1 || i == 3 { offset + 1 } else { offset + 3 };
 
-                offset += input_offset;
-                // Since the new bits are all zero, we have to set the hash of 0 as the respective output
-                zero_state.iter().enumerate().for_each(|(j, &value)| {
-                    let block_offset = j * BLOCK_SIZE;
-                    let block_pos = offset + block_offset;
-                    for k in 0..32 {
-                        let bit = ((value >> k) & 1) as u64;
-                        let bit_pos = k % BITS_IN_PARALLEL;
-                        let bit_offset = (k - bit_pos) * NUM_SHA256F_PER_CIRCUIT / BITS_IN_PARALLEL;
-                        let pos = block_pos + bit_offset;
-                        for w in rem_inputs..NUM_SHA256F_PER_CIRCUIT {
-                            update_bit_val(trace, pos + w, bit, w, bit_pos, w == 0);
-                        }
-                    }
-                });
-            }
-
-            // Fill the remaining circuits with the hash of 0
-            let next_circuit = num_inputs.div_ceil(NUM_SHA256F_PER_CIRCUIT);
-            zero_state.iter().enumerate().for_each(|(j, &value)| {
-                for s in next_circuit..self.num_available_circuits {
-                    let circuit_offset = s * self.circuit_size;
-                    let block_offset = j * BLOCK_SIZE;
-                    for k in 0..32 {
-                        let bit = ((value >> k) & 1) as u64;
-                        let bit_pos = k % BITS_IN_PARALLEL;
-                        let bit_offset = (k - bit_pos) * NUM_SHA256F_PER_CIRCUIT / BITS_IN_PARALLEL;
-                        let pos = initial_offset
-                            + circuit_offset
-                            + input_offset
-                            + block_offset
-                            + bit_offset;
-                        for w in 0..NUM_SHA256F_PER_CIRCUIT {
-                            update_bit_val(trace, pos + w, bit, w, bit_pos, w == 0);
-                        }
-                    }
-                }
-            });
-        }
-
-        // 2] Set the values of free_in_a, free_in_b, free_in_c and free_in_d using the script
-
-        // Divide input bits between state bits and hash input bits
-        let state_bits: Vec<[u64; STATE_SIZE_BITS]> =
-            inputs_bits.iter().map(|bits| bits[..STATE_SIZE_BITS].try_into().unwrap()).collect();
-        let hash_input_bits: Vec<[u64; INPUT_SIZE_BITS]> =
-            inputs_bits.iter().map(|bits| bits[STATE_SIZE_BITS..].try_into().unwrap()).collect();
-
-        let trace_rows = trace.row_slice_mut();
-
-        let row0 = trace_rows[0];
-
-        let mut trace_slice = &mut trace_rows[1..];
-
-        let mut par_traces = Vec::new();
-
-        for _ in 0..self.num_available_circuits {
-            let take = self.circuit_size.min(trace_slice.len());
-            let (head, tail) = trace_slice.split_at_mut(take);
-            par_traces.push(head);
-            trace_slice = tail;
-        }
-
-        let program = &self.script.program;
-        par_traces.into_par_iter().enumerate().for_each(|(i, par_trace)| {
-            for line in program.iter() {
-                let row = line.ref_ - 1;
-
-                let a_val = get_val(par_trace, &row0, &state_bits, &hash_input_bits, i, &line.in1);
-                set_col(par_trace, |row| &mut row.free_in_a, row, a_val);
-
-                let b_val = get_val(par_trace, &row0, &state_bits, &hash_input_bits, i, &line.in2);
-                set_col(par_trace, |row| &mut row.free_in_b, row, b_val);
-
-                if let Some(in3) = &line.in3 {
-                    let c_val = get_val(par_trace, &row0, &state_bits, &hash_input_bits, i, in3);
-                    set_col(par_trace, |row| &mut row.free_in_c, row, c_val);
-                }
-                let c_val = get_col(par_trace, |row| &row.free_in_c, row);
-
-                let op = &line.op;
-                let d_val;
-                let op_val;
-                if op == "xor" {
-                    d_val = a_val ^ b_val ^ c_val;
-                    op_val = Sha256fTableGateOp::Xor;
-                } else if op == "ch" {
-                    d_val = (a_val & b_val) ^ ((a_val ^ MASK_CHUNK_BITS_SHA256F) & c_val);
-                    op_val = Sha256fTableGateOp::Ch;
-                } else if op == "maj" {
-                    d_val = (a_val & b_val) ^ (a_val & c_val) ^ (b_val & c_val);
-                    op_val = Sha256fTableGateOp::Maj;
-                } else if op == "add" {
-                    d_val = a_val ^ b_val ^ c_val;
-                    op_val = Sha256fTableGateOp::Add;
-
-                    // Compute and set the carry
-                    let carry = (a_val & b_val) | (a_val & c_val) | (b_val & c_val);
-                    set_col(par_trace, |row| &mut row.carry, row, carry);
-                    set_col(par_trace, |row| &mut row.free_in_c, row + 1, carry);
+            // Locate the state bits in the trace
+            let is_a = i < 2;
+            for j in 0..32 {
+                let bit = ((word_high >> j) & 1) as u8;
+                if is_a {
+                    trace[row].a[j] = F::from_u8(bit);
                 } else {
-                    panic!("Invalid operation: {}", op);
-                }
-
-                set_col(par_trace, |row| &mut row.free_in_d, row, d_val);
-
-                // Update the multiplicity table for the circuit
-                for j in 0..CHUNKS_SHA256F {
-                    let a = (a_val >> (j * BITS_SHA256F)) & MASK_BITS_SHA256F;
-                    let b = (b_val >> (j * BITS_SHA256F)) & MASK_BITS_SHA256F;
-                    let c = (c_val >> (j * BITS_SHA256F)) & MASK_BITS_SHA256F;
-                    let table_row = Sha256fTableSM::calculate_table_row(&op_val, a, b, c);
-                    self.sha256f_table_sm.update_input(table_row, 1);
+                    trace[row].e[j] = F::from_u8(bit);
                 }
             }
-        });
+            row -= 1;
+            for j in 0..32 {
+                let bit = ((word_low >> j) & 1) as u8;
+                if is_a {
+                    trace[row].a[j] = F::from_u8(bit);
+                } else {
+                    trace[row].e[j] = F::from_u8(bit);
+                }
+            }
+        }
+        offset += CLOCKS_LOAD_STATE;
 
-        fn update_bit_val<F: PrimeField64>(
-            trace: &mut Sha256fTrace<F>,
-            pos: usize,
-            bit: u64,
-            circuit_pos: usize,
-            bit_pos: usize,
-            reset: bool,
-        ) {
-            trace[pos].bit[bit_pos] = F::from_u64(bit);
-            trace[pos + 1].val[bit_pos] = if reset {
-                F::from_u64(bit << circuit_pos)
+        // Compute the load input stage
+        let mut w = [0u32; 16];
+        for i in 0..CLOCKS_LOAD_INPUT {
+            let word = input[i / 2];
+
+            // Store the input as u32 for further processing
+            w[i] = if i % 2 == 0 { (word >> 32) as u32 } else { (word & 0xFFFF_FFFF) as u32 };
+
+            // Compute the a and e values for the current input
+            let [old_a, old_b, old_c, old_d, old_e, old_f, old_g, old_h] = prev_state;
+            let (a, e) =
+                compute_ae(old_a, old_b, old_c, old_d, old_e, old_f, old_g, old_h, w[i], RC[i]);
+
+            let (a_carry, a) = ((a >> 32) as u8, (a & 0xFFFF_FFFF) as u32);
+            let (e_carry, e) = ((e >> 32) as u8, (e & 0xFFFF_FFFF) as u32);
+
+            let row = offset + i;
+
+            // Locate the carry
+            trace[row].new_a_carry_bits = F::from_u8(a_carry);
+            trace[row].new_e_carry_bits = F::from_u8(e_carry);
+            self.std.range_check(a_carry as i64, 1, self.a_range_id);
+            self.std.range_check(e_carry as i64, 1, self.e_range_id);
+
+            // Locate the input bits in the trace
+            for j in 0..32 {
+                let bit_a = ((a >> j) & 1) as u8;
+                let bit_e = ((e >> j) & 1) as u8;
+                let bit_w = ((w[i] >> j) & 1) as u8;
+                trace[row].a[j] = F::from_u8(bit_a);
+                trace[row].e[j] = F::from_u8(bit_e);
+                trace[row].w[j] = F::from_u8(bit_w);
+            }
+
+            // Update prev_state for the next iteration
+            prev_state[7] = old_g;
+            prev_state[6] = old_f;
+            prev_state[5] = old_e;
+            prev_state[4] = e;
+            prev_state[3] = old_c;
+            prev_state[2] = old_b;
+            prev_state[1] = old_a;
+            prev_state[0] = a;
+        }
+        offset += CLOCKS_LOAD_INPUT;
+
+        // Compute the mixing stage
+        for i in 0..CLOCKS_MIXING {
+            let [old_w2, old_w7, old_w15, old_w16] = [
+                w[CLOCKS_LOAD_INPUT - 2],
+                w[CLOCKS_LOAD_INPUT - 7],
+                w[CLOCKS_LOAD_INPUT - 15],
+                w[CLOCKS_LOAD_INPUT - 16],
+            ];
+            let new_w = compute_w(old_w2, old_w7, old_w15, old_w16);
+            let (new_w_carry, new_w) = ((new_w >> 32) as u8, (new_w & 0xFFFF_FFFF) as u32);
+
+            let [old_a, old_b, old_c, old_d, old_e, old_f, old_g, old_h] = prev_state;
+            #[rustfmt::skip]
+            let (a, e) = compute_ae(old_a, old_b, old_c, old_d, old_e, old_f, old_g, old_h, new_w, RC[CLOCKS_LOAD_INPUT + i]);
+
+            let (a_carry, a) = ((a >> 32) as u8, (a & 0xFFFF_FFFF) as u32);
+            let (e_carry, e) = ((e >> 32) as u8, (e & 0xFFFF_FFFF) as u32);
+
+            let row = offset + i;
+
+            // Locate the carry
+            trace[row].new_a_carry_bits = F::from_u8(a_carry);
+            trace[row].new_e_carry_bits = F::from_u8(e_carry);
+            trace[row].new_w_carry_bits = F::from_u8(new_w_carry);
+            self.std.range_check(a_carry as i64, 1, self.a_range_id);
+            self.std.range_check(e_carry as i64, 1, self.e_range_id);
+
+            for j in 0..32 {
+                let bit_a = ((a >> j) & 1) as u8;
+                let bit_e = ((e >> j) & 1) as u8;
+                let bit_w = ((new_w >> j) & 1) as u8;
+                trace[row].a[j] = F::from_u8(bit_a);
+                trace[row].e[j] = F::from_u8(bit_e);
+                trace[row].w[j] = F::from_u8(bit_w);
+            }
+
+            // Update prev_state for the next iteration
+            prev_state[7] = old_g;
+            prev_state[6] = old_f;
+            prev_state[5] = old_e;
+            prev_state[4] = e;
+            prev_state[3] = old_c;
+            prev_state[2] = old_b;
+            prev_state[1] = old_a;
+            prev_state[0] = a;
+
+            // Update the w array for the next iteration
+            for j in 0..15 {
+                w[j] = w[j + 1];
+            }
+            w[15] = new_w;
+        }
+        offset += CLOCKS_MIXING;
+
+        for i in 0..CLOCKS_WRITE_STATE {
+            let prev = state[i];
+            let prev_high = prev >> 32;
+            let prev_low = prev & 0xFFFF_FFFF;
+
+            let curr_high = (prev_state[2 * i]) as u64;
+            let curr_low = (prev_state[2 * i + 1]) as u64;
+
+            let new_high = curr_high + prev_high;
+            let new_low = curr_low + prev_low;
+            let (new_high_carry, new_high) =
+                ((new_high >> 32) as u8, (new_high & 0xFFFF_FFFF) as u32);
+            let (new_low_carry, new_low) = ((new_low >> 32) as u8, (new_low & 0xFFFF_FFFF) as u32);
+
+            let mut row = if i == 1 || i == 3 { offset + 1 } else { offset + 3 };
+
+            // Locate the state bits in the trace
+            let is_a = i < 2;
+            if is_a {
+                trace[row].new_a_carry_bits = F::from_u8(new_high_carry);
+                self.std.range_check(new_high_carry as i64, 1, self.a_range_id);
             } else {
-                trace[pos].val[bit_pos] + F::from_u64(bit << circuit_pos)
-            };
-        }
+                trace[row].new_e_carry_bits = F::from_u8(new_high_carry);
+                self.std.range_check(new_high_carry as i64, 1, self.e_range_id);
+            }
 
-        fn get_val<F: PrimeField64>(
-            trace: &[Sha256fTraceRow<F>],
-            row0: &Sha256fTraceRow<F>,
-            state_bits: &[[u64; STATE_SIZE_BITS]],
-            hash_input_bits: &[[u64; INPUT_SIZE_BITS]],
-            circuit: usize,
-            gate_input: &InputType,
-        ) -> u64 {
-            match gate_input {
-                InputType::Wired { gate, pin, .. } => match pin.as_str() {
-                    "in1" => {
-                        if *gate > 0 {
-                            get_col(trace, |row| &row.free_in_a, *gate - 1)
-                        } else {
-                            get_col_row(row0, |row| &row.free_in_a)
-                        }
-                    }
-                    "in2" => {
-                        if *gate > 0 {
-                            get_col(trace, |row| &row.free_in_b, *gate - 1)
-                        } else {
-                            get_col_row(row0, |row| &row.free_in_b)
-                        }
-                    }
-                    "in3" => {
-                        if *gate > 0 {
-                            get_col(trace, |row| &row.free_in_c, *gate - 1)
-                        } else {
-                            get_col_row(row0, |row| &row.free_in_c)
-                        }
-                    }
-                    "out" => {
-                        if *gate > 0 {
-                            get_col(trace, |row| &row.free_in_d, *gate - 1)
-                        } else {
-                            get_col_row(row0, |row| &row.free_in_d)
-                        }
-                    }
-                    _ => panic!("Invalid pin: {}", pin),
-                },
-                InputType::Input { bit, .. } => hash_input_bits[circuit][*bit],
-                InputType::InputState { bit, .. } => state_bits[circuit][*bit],
+            for j in 0..32 {
+                let bit = ((new_high >> j) & 1) as u8;
+                if is_a {
+                    trace[row].a[j] = F::from_u8(bit);
+                } else {
+                    trace[row].e[j] = F::from_u8(bit);
+                }
+            }
+            row -= 1;
+
+            if is_a {
+                trace[row].new_a_carry_bits = F::from_u8(new_low_carry);
+                self.std.range_check(new_low_carry as i64, 1, self.a_range_id);
+            } else {
+                trace[row].new_e_carry_bits = F::from_u8(new_low_carry);
+                self.std.range_check(new_low_carry as i64, 1, self.e_range_id);
+            }
+
+            for j in 0..32 {
+                let bit = ((new_low >> j) & 1) as u8;
+                if is_a {
+                    trace[row].a[j] = F::from_u8(bit);
+                } else {
+                    trace[row].e[j] = F::from_u8(bit);
+                }
             }
         }
 
-        fn set_col<F: PrimeField64>(
-            trace: &mut [Sha256fTraceRow<F>],
-            cols: impl Fn(&mut Sha256fTraceRow<F>) -> &mut [F; CHUNKS_SHA256F],
-            index: usize,
-            value: u64,
-        ) {
-            let mut _value = value;
-            let row = &mut trace[index];
-            let cols = cols(row);
-            for col in cols.iter_mut() {
-                *col = F::from_u64(_value & MASK_BITS_SHA256F);
-                _value >>= BITS_SHA256F;
-            }
+        // Perform the zero range checks
+        self.std.range_check(0, CLOCKS_LOAD_STATE as u64, self.a_range_id);
+        self.std.range_check(0, CLOCKS_LOAD_STATE as u64, self.e_range_id);
+
+        #[rustfmt::skip]
+        #[allow(clippy::too_many_arguments)]
+        fn compute_ae(old_a: u32, old_b: u32, old_c: u32, old_d: u32, old_e: u32, old_f: u32, old_g: u32, old_h: u32, w: u32, k: u32) -> (u64, u64) {
+            let s0 = rotate_right(old_a, 2) ^ rotate_right(old_a, 13) ^ rotate_right(old_a, 22);
+            let s1 = rotate_right(old_e, 6) ^ rotate_right(old_e, 11) ^ rotate_right(old_e, 25);
+            let t1 = (old_h as u64) + (s1 as u64) + (ch(old_e, old_f, old_g) as u64) + (k as u64) + (w as u64);
+            let t2 = (s0 as u64) + (maj(old_a, old_b, old_c) as u64);
+            let a = (t1 as u64) + (t2 as u64);
+            let e = (old_d as u64) + (t1 as u64);
+            (a, e)
+            // (s0 as u64, s1 as u64)
         }
 
-        fn get_col<F: PrimeField64>(
-            trace: &[Sha256fTraceRow<F>],
-            cols: impl Fn(&Sha256fTraceRow<F>) -> &[F; CHUNKS_SHA256F],
-            index: usize,
-        ) -> u64 {
-            let mut value = 0;
-            let row = &trace[index];
-            let cols = cols(row);
-            for (i, col) in cols.iter().enumerate() {
-                let col_i_val = F::as_canonical_u64(col);
-                value += col_i_val << ((i * BITS_SHA256F) as u64);
-            }
-            value
+        fn compute_w(old_w2: u32, old_w7: u32, old_w15: u32, old_w16: u32) -> u64 {
+            let s0 = rotate_right(old_w15, 7) ^ rotate_right(old_w15, 18) ^ shift_right(old_w15, 3);
+            let s1 = rotate_right(old_w2, 17) ^ rotate_right(old_w2, 19) ^ shift_right(old_w2, 10);
+            (s1 as u64) + (old_w7 as u64) + (s0 as u64) + (old_w16 as u64)
         }
 
-        fn get_col_row<F: PrimeField64>(
-            trace_row: &Sha256fTraceRow<F>,
-            cols: impl Fn(&Sha256fTraceRow<F>) -> &[F; CHUNKS_SHA256F],
-        ) -> u64 {
-            let mut value = 0;
-            let row = trace_row;
-            let cols = cols(row);
-            for (i, col) in cols.iter().enumerate() {
-                let col_i_val = F::as_canonical_u64(col);
-                value += col_i_val << ((i * BITS_SHA256F) as u64);
-            }
-            value
+        fn rotate_right(x: u32, n: u32) -> u32 {
+            (x >> n) | (x << (32 - n))
+        }
+
+        fn shift_right(x: u32, n: u32) -> u32 {
+            x >> n
+        }
+
+        fn maj(x: u32, y: u32, z: u32) -> u32 {
+            (x & y) ^ (x & z) ^ (y & z)
+        }
+
+        fn ch(x: u32, y: u32, z: u32) -> u32 {
+            (x & y) ^ (!x & z)
         }
     }
 
     /// Computes the witness for a series of inputs and produces an `AirInstance`.
     ///
     /// # Arguments
+    /// * `sctx` - The setup context containing the setup data.
     /// * `inputs` - A slice of operations to process.
     ///
     /// # Returns
     /// An `AirInstance` containing the computed witness data.
-    pub fn compute_witness(
-        &self,
-        inputs: &[Vec<Sha256fInput>],
-        trace_buffer: Vec<F>,
-    ) -> AirInstance<F> {
-        timer_start_trace!(SHA256F_TRACE);
-        let mut sha256f_trace = Sha256fTrace::new_from_vec_zeroes(trace_buffer);
+    pub fn compute_witness(&self, inputs: &[Vec<Sha256fInput>], trace_buffer: Vec<F>,) -> AirInstance<F> {
+        let mut sha256f_trace = Sha256fTrace::new_from_vec(trace_buffer);
         let num_rows = sha256f_trace.num_rows();
+        let num_available_sha256fs = self.num_available_sha256fs;
 
         // Check that we can fit all the sha256fs in the trace
         let num_inputs = inputs.iter().map(|v| v.len()).sum::<usize>();
-        let num_circuits_needed = num_inputs.div_ceil(NUM_SHA256F_PER_CIRCUIT);
-        let num_rows_constants = 1; // Number of rows used for the constants
-        let num_padding_rows = (num_rows - num_rows_constants) % self.circuit_size;
-        let num_rows_needed =
-            num_rows_constants + num_circuits_needed * self.circuit_size + num_padding_rows;
-
-        // Sanity checks
-        assert!(
-            num_inputs <= self.num_available_sha256fs,
-            "Exceeded available Sha256fs inputs: requested {}, but only {} are available.",
-            num_inputs,
-            self.num_available_sha256fs
-        );
-        assert!(num_circuits_needed <= self.num_available_circuits);
-        assert!(num_rows_needed <= num_rows);
+        let num_rows_filled = num_inputs * CLOCKS;
+        let num_rows_needed = if num_inputs < num_available_sha256fs {
+            num_inputs * CLOCKS
+        } else if num_inputs == num_available_sha256fs {
+            num_rows
+        } else {
+            panic!(
+                "Exceeded available Sha256fs inputs: requested {}, but only {} are available.",
+                num_inputs, self.num_available_sha256fs
+            );
+        };
 
         tracing::info!(
             "··· Creating Sha256f instance [{} / {} rows filled {:.2}%]",
@@ -524,52 +338,94 @@ impl<F: PrimeField64> Sha256fSM<F> {
             num_rows_needed as f64 / num_rows as f64 * 100.0
         );
 
-        // Set a = 0b00..00, b = 0b11..11 and c = 0b00..00 at the first row
-        // Set, e.g., the operation to be an XOR and set d = 0b11..11 = b = a ^ b ^ c
-        let mut row: Sha256fTraceRow<F> = Default::default();
-        let zeros = 0u64;
-        let ones = MASK_BITS_SHA256F;
-        let gate_op = self.sha256f_fixed[0].GATE_OP.as_canonical_u64();
-        // Sanity check
-        assert_eq!(gate_op, Sha256fTableGateOp::Xor as u64, "Invalid first row gate operation");
-        for i in 0..CHUNKS_SHA256F {
-            row.free_in_a[i] = F::from_u64(zeros);
-            row.free_in_b[i] = F::from_u64(ones);
-            row.free_in_c[i] = F::from_u64(zeros);
-            row.free_in_d[i] = F::from_u64(ones);
+        timer_start_trace!(SHA256F_TRACE);
+        let mut trace_rows = sha256f_trace.row_slice_mut();
+        let mut par_traces = Vec::new();
+        let mut inputs_indexes = Vec::new();
+        for (i, inputs) in inputs.iter().enumerate() {
+            for (j, _) in inputs.iter().enumerate() {
+                let (head, tail) = trace_rows.split_at_mut(CLOCKS);
+                par_traces.push(head);
+                inputs_indexes.push((i, j));
+                trace_rows = tail;
+            }
         }
 
-        // Assign the first row
-        sha256f_trace[0] = row;
-
-        // Update the multiplicity table
-        let table_row =
-            Sha256fTableSM::calculate_table_row(&Sha256fTableGateOp::Xor, zeros, ones, zeros);
-        self.sha256f_table_sm.update_input(table_row, CHUNKS_SHA256F as u64);
-
-        // Fill the rest of the trace
-        // Flatten all the inputs, since I need to process them at least in chunks of NUM_SHA256F_PER_CIRCUIT
-        let inputs = inputs.iter().flatten();
-        self.process_trace(&mut sha256f_trace, num_rows_constants, inputs, num_inputs);
+        // Fill the trace
+        par_traces.into_par_iter().enumerate().for_each(|(index, trace)| {
+            let input_index = inputs_indexes[index];
+            let input = &inputs[input_index.0][input_index.1];
+            self.process_input(input, trace);
+        });
         timer_stop_and_log_trace!(SHA256F_TRACE);
 
         timer_start_trace!(SHA256F_PADDING);
-        // A row with all zeros satisfies the constraints (assuming the operation to be XOR(0,0,0)=0)
-        let padding_row: Sha256fTraceRow<F> = Default::default();
-        for i in (num_rows_constants + self.circuit_size * self.num_available_circuits)..num_rows {
-            let gate_op = self.sha256f_fixed[i].GATE_OP.as_canonical_u64();
-            // Sanity check
-            assert_eq!(
-                gate_op,
-                Sha256fTableGateOp::Xor as u64,
-                "Invalid padding dummy gate operation"
+        // Set a = e = w = 0 for the state and input rows
+        let zero_row = Sha256fTraceRow::<F>::default();
+
+        // precompute compute_ae() with initial a = e = 0 (PC_A and PC_E)
+        // compute_w() with w = 0 is equal to 0, nothing to do
+        let mut mid_rows = [Sha256fTraceRow::<F>::default(); 64];
+        for i in 0..64 {
+            let a = PC_A[i];
+            let e = PC_E[i];
+            let (a_carry, a) = ((a >> 32) as u8, (a & 0xFFFF_FFFF) as u32);
+            let (e_carry, e) = ((e >> 32) as u8, (e & 0xFFFF_FFFF) as u32);
+            mid_rows[i].new_a_carry_bits = F::from_u8(a_carry);
+            mid_rows[i].new_e_carry_bits = F::from_u8(e_carry);
+            for j in 0..32 {
+                let bit_a = ((a >> j) & 1) as u8;
+                let bit_e = ((e >> j) & 1) as u8;
+                mid_rows[i].a[j] = F::from_u8(bit_a);
+                mid_rows[i].e[j] = F::from_u8(bit_e);
+            }
+
+            self.std.range_check(
+                a_carry as i64,
+                (num_available_sha256fs - num_inputs) as u64,
+                self.a_range_id,
             );
-
-            let table_row = Sha256fTableSM::calculate_table_row(&Sha256fTableGateOp::Xor, 0, 0, 0);
-            self.sha256f_table_sm.update_input(table_row, CHUNKS_SHA256F as u64);
-
-            sha256f_trace[i] = padding_row;
+            self.std.range_check(
+                e_carry as i64,
+                (num_available_sha256fs - num_inputs) as u64,
+                self.e_range_id,
+            );
         }
+
+        // At the end, we should have that a === 4'and e === 4'e
+        let mut final_rows = [Sha256fTraceRow::<F>::default(); 4];
+        for i in 0..4 {
+            let a = (PC_A[60 + i] & 0xFFFF_FFFF) as u32;
+            let e = (PC_E[60 + i] & 0xFFFF_FFFF) as u32;
+            for j in 0..32 {
+                let bit_a = ((a >> j) & 1) as u8;
+                let bit_e = ((e >> j) & 1) as u8;
+                final_rows[i].a[j] = F::from_u8(bit_a);
+                final_rows[i].e[j] = F::from_u8(bit_e);
+            }
+        }
+
+        const CLOCKS_OP: usize = CLOCKS_LOAD_STATE + CLOCKS_LOAD_INPUT + CLOCKS_MIXING;
+        // The last (CLOCKS + NUM_NON_USABLE_ROWS) have CLK_0 desactivated, so
+        // a trace full of zeroes passes the constraints
+        for row in num_rows_filled..(num_rows - self.num_non_usable_rows - CLOCKS) {
+            let row_r = row % CLOCKS;
+            sha256f_trace[row] = if row_r < CLOCKS_LOAD_STATE {
+                zero_row
+            } else if row_r < CLOCKS_OP {
+                mid_rows[row_r - CLOCKS_LOAD_STATE]
+            } else {
+                final_rows[row_r - CLOCKS_OP]
+            };
+        }
+
+        // Perform the zero range checks
+        let count_zeros = (num_available_sha256fs - num_inputs)
+            * (CLOCKS_LOAD_STATE + CLOCKS_WRITE_STATE)
+            + CLOCKS
+            + self.num_non_usable_rows;
+        self.std.range_check(0, count_zeros as u64, self.a_range_id);
+        self.std.range_check(0, count_zeros as u64, self.e_range_id);
         timer_stop_and_log_trace!(SHA256F_PADDING);
 
         AirInstance::new_from_trace(FromTrace::new(&mut sha256f_trace))
