--- precompiles/sha256f_direct/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "precomp-sha256f"
-version = { workspace = true }
-edition = { workspace = true }
-license = { workspace = true }
-keywords = { workspace = true }
-repository = { workspace = true }
-categories = { workspace = true }
-
-[dependencies]
-zisk-core = { path = "../../core" }
-zisk-common = { path = "../../common" }
-data-bus = { path = "../../data-bus" }
-zisk-pil = { path = "../../pil" }
-precompiles-common = { path = "../common" }
-circuit = { path = "../../tools/circuit" }
-precompiles-helpers = { path = "../helpers" }
-
-proofman = { workspace = true }
-proofman-common = { workspace = true }
-proofman-macros = { workspace = true }
-proofman-util = { workspace = true }
-witness = { workspace = true }
-pil-std-lib = { workspace = true }
-
-fields = { workspace=true }
-tracing = { workspace = true }
-rayon = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-
-clap = "4.0"
-sha2 = { version = "0.10.9", features = ["compress"] }
-generic-array = "0.14"
-
-[features]
-default = []
-no_lib_link = ["proofman-common/no_lib_link"]
-diagnostic = ["proofman-macros/diagnostic", "proofman/diagnostic"]
-disable_distributed = ["proofman/disable_distributed", "proofman-common/disable_distributed"]
