--- precompiles/sha256f_direct/src/lib.rs
@@ -1,17 +1,0 @@
-mod sha256f;
-mod sha256f_bus_device;
-mod sha256f_constants;
-mod sha256f_gen_mem_inputs;
-mod sha256f_input;
-mod sha256f_instance;
-mod sha256f_manager;
-mod sha256f_planner;
-
-pub use sha256f::*;
-pub use sha256f_bus_device::*;
-pub use sha256f_constants::*;
-pub use sha256f_gen_mem_inputs::*;
-pub use sha256f_input::*;
-pub use sha256f_instance::*;
-pub use sha256f_manager::*;
-pub use sha256f_planner::*;
