--- precompiles/sha256f_direct/src/sha256f.rs
@@ -1,434 +1,0 @@
-use core::panic;
-use std::sync::Arc;
-
-use fields::PrimeField64;
-use rayon::prelude::*;
-
-use pil_std_lib::Std;
-use proofman_common::{AirInstance, FromTrace};
-use proofman_util::{timer_start_trace, timer_stop_and_log_trace};
-use zisk_pil::{Sha256fTrace, Sha256fTraceRow};
-
-use super::{sha256f_constants::*, Sha256fInput};
-
-/// The `Sha256fSM` struct encapsulates the logic of the Sha256f State Machine.
-pub struct Sha256fSM<F: PrimeField64> {
-    /// Reference to the PIL2 standard library.
-    pub std: Arc<Std<F>>,
-
-    /// Number of available sha256fs in the trace.
-    pub num_available_sha256fs: usize,
-
-    num_non_usable_rows: usize,
-
-    /// Range checks ID's
-    a_range_id: usize,
-    e_range_id: usize,
-}
-
-impl<F: PrimeField64> Sha256fSM<F> {
-    /// Creates a new Sha256f State Machine instance.
-    ///
-    /// # Returns
-    /// A new `Sha256fSM` instance.
-    pub fn new(std: Arc<Std<F>>) -> Arc<Self> {
-        // Compute some useful values
-        let num_available_sha256fs = Sha256fTrace::<usize>::NUM_ROWS / CLOCKS - 1;
-        let num_non_usable_rows = Sha256fTrace::<usize>::NUM_ROWS % CLOCKS;
-
-        let a_range_id = std.get_range(0, (1 << 3) - 1, None);
-        let e_range_id = std.get_range(0, (1 << 3) - 1, None);
-
-        Arc::new(Self { std, num_available_sha256fs, num_non_usable_rows, a_range_id, e_range_id })
-    }
-
-    /// Processes a slice of operation data, updating the trace and multiplicities.
-    ///
-    /// # Arguments
-    /// * `trace` - A mutable reference to the Sha256f trace.
-    /// * `num_circuits` - The number of circuits to process.
-    /// * `input` - The operation data to process.
-    /// * `multiplicity` - A mutable slice to update with multiplicities for the operation.
-    #[inline(always)]
-    pub fn process_input(&self, input: &Sha256fInput, trace: &mut [Sha256fTraceRow<F>]) {
-        let step_main = input.step_main;
-        let addr_main = input.addr_main;
-        let state_addr = input.state_addr;
-        let input_addr = input.input_addr;
-        let state = &input.state;
-        let input = &input.input;
-
-        // Fill the step_addr
-        trace[0].step_addr = F::from_u64(step_main); // STEP_MAIN
-        trace[1].step_addr = F::from_u32(addr_main); // ADDR_OP
-        trace[2].step_addr = F::from_u32(state_addr); // ADDR_STATE
-        trace[3].step_addr = F::from_u32(input_addr); // ADDR_INPUT
-        trace[4].step_addr = F::from_u32(state_addr); // ADDR_IND_0
-        trace[5].step_addr = F::from_u32(input_addr); // ADDR_IND_1
-
-        // Activate the clk_0 selector
-        trace[0].in_use_clk_0 = F::ONE;
-
-        // Activate the in_use selector
-        for r in trace.iter_mut().take(18) {
-            r.in_use = F::ONE;
-        }
-
-        // Compute the load state stage
-        let mut offset = 0;
-        let mut prev_state = [0u32; 8];
-        for i in 0..CLOCKS_LOAD_STATE {
-            let word = state[i];
-            let word_high = (word >> 32) as u32;
-            let word_low = (word & 0xFFFF_FFFF) as u32;
-
-            // Store the state as u32 for further processing
-            prev_state[2 * i] = word_high;
-            prev_state[2 * i + 1] = word_low;
-
-            let mut row = if i == 1 || i == 3 { offset + 1 } else { offset + 3 };
-
-            // Locate the state bits in the trace
-            let is_a = i < 2;
-            for j in 0..32 {
-                let bit = ((word_high >> j) & 1) as u8;
-                if is_a {
-                    trace[row].a[j] = F::from_u8(bit);
-                } else {
-                    trace[row].e[j] = F::from_u8(bit);
-                }
-            }
-            row -= 1;
-            for j in 0..32 {
-                let bit = ((word_low >> j) & 1) as u8;
-                if is_a {
-                    trace[row].a[j] = F::from_u8(bit);
-                } else {
-                    trace[row].e[j] = F::from_u8(bit);
-                }
-            }
-        }
-        offset += CLOCKS_LOAD_STATE;
-
-        // Compute the load input stage
-        let mut w = [0u32; 16];
-        for i in 0..CLOCKS_LOAD_INPUT {
-            let word = input[i / 2];
-
-            // Store the input as u32 for further processing
-            w[i] = if i % 2 == 0 { (word >> 32) as u32 } else { (word & 0xFFFF_FFFF) as u32 };
-
-            // Compute the a and e values for the current input
-            let [old_a, old_b, old_c, old_d, old_e, old_f, old_g, old_h] = prev_state;
-            let (a, e) =
-                compute_ae(old_a, old_b, old_c, old_d, old_e, old_f, old_g, old_h, w[i], RC[i]);
-
-            let (a_carry, a) = ((a >> 32) as u8, (a & 0xFFFF_FFFF) as u32);
-            let (e_carry, e) = ((e >> 32) as u8, (e & 0xFFFF_FFFF) as u32);
-
-            let row = offset + i;
-
-            // Locate the carry
-            trace[row].new_a_carry_bits = F::from_u8(a_carry);
-            trace[row].new_e_carry_bits = F::from_u8(e_carry);
-            self.std.range_check(a_carry as i64, 1, self.a_range_id);
-            self.std.range_check(e_carry as i64, 1, self.e_range_id);
-
-            // Locate the input bits in the trace
-            for j in 0..32 {
-                let bit_a = ((a >> j) & 1) as u8;
-                let bit_e = ((e >> j) & 1) as u8;
-                let bit_w = ((w[i] >> j) & 1) as u8;
-                trace[row].a[j] = F::from_u8(bit_a);
-                trace[row].e[j] = F::from_u8(bit_e);
-                trace[row].w[j] = F::from_u8(bit_w);
-            }
-
-            // Update prev_state for the next iteration
-            prev_state[7] = old_g;
-            prev_state[6] = old_f;
-            prev_state[5] = old_e;
-            prev_state[4] = e;
-            prev_state[3] = old_c;
-            prev_state[2] = old_b;
-            prev_state[1] = old_a;
-            prev_state[0] = a;
-        }
-        offset += CLOCKS_LOAD_INPUT;
-
-        // Compute the mixing stage
-        for i in 0..CLOCKS_MIXING {
-            let [old_w2, old_w7, old_w15, old_w16] = [
-                w[CLOCKS_LOAD_INPUT - 2],
-                w[CLOCKS_LOAD_INPUT - 7],
-                w[CLOCKS_LOAD_INPUT - 15],
-                w[CLOCKS_LOAD_INPUT - 16],
-            ];
-            let new_w = compute_w(old_w2, old_w7, old_w15, old_w16);
-            let (new_w_carry, new_w) = ((new_w >> 32) as u8, (new_w & 0xFFFF_FFFF) as u32);
-
-            let [old_a, old_b, old_c, old_d, old_e, old_f, old_g, old_h] = prev_state;
-            #[rustfmt::skip]
-            let (a, e) = compute_ae(old_a, old_b, old_c, old_d, old_e, old_f, old_g, old_h, new_w, RC[CLOCKS_LOAD_INPUT + i]);
-
-            let (a_carry, a) = ((a >> 32) as u8, (a & 0xFFFF_FFFF) as u32);
-            let (e_carry, e) = ((e >> 32) as u8, (e & 0xFFFF_FFFF) as u32);
-
-            let row = offset + i;
-
-            // Locate the carry
-            trace[row].new_a_carry_bits = F::from_u8(a_carry);
-            trace[row].new_e_carry_bits = F::from_u8(e_carry);
-            trace[row].new_w_carry_bits = F::from_u8(new_w_carry);
-            self.std.range_check(a_carry as i64, 1, self.a_range_id);
-            self.std.range_check(e_carry as i64, 1, self.e_range_id);
-
-            for j in 0..32 {
-                let bit_a = ((a >> j) & 1) as u8;
-                let bit_e = ((e >> j) & 1) as u8;
-                let bit_w = ((new_w >> j) & 1) as u8;
-                trace[row].a[j] = F::from_u8(bit_a);
-                trace[row].e[j] = F::from_u8(bit_e);
-                trace[row].w[j] = F::from_u8(bit_w);
-            }
-
-            // Update prev_state for the next iteration
-            prev_state[7] = old_g;
-            prev_state[6] = old_f;
-            prev_state[5] = old_e;
-            prev_state[4] = e;
-            prev_state[3] = old_c;
-            prev_state[2] = old_b;
-            prev_state[1] = old_a;
-            prev_state[0] = a;
-
-            // Update the w array for the next iteration
-            for j in 0..15 {
-                w[j] = w[j + 1];
-            }
-            w[15] = new_w;
-        }
-        offset += CLOCKS_MIXING;
-
-        for i in 0..CLOCKS_WRITE_STATE {
-            let prev = state[i];
-            let prev_high = prev >> 32;
-            let prev_low = prev & 0xFFFF_FFFF;
-
-            let curr_high = (prev_state[2 * i]) as u64;
-            let curr_low = (prev_state[2 * i + 1]) as u64;
-
-            let new_high = curr_high + prev_high;
-            let new_low = curr_low + prev_low;
-            let (new_high_carry, new_high) =
-                ((new_high >> 32) as u8, (new_high & 0xFFFF_FFFF) as u32);
-            let (new_low_carry, new_low) = ((new_low >> 32) as u8, (new_low & 0xFFFF_FFFF) as u32);
-
-            let mut row = if i == 1 || i == 3 { offset + 1 } else { offset + 3 };
-
-            // Locate the state bits in the trace
-            let is_a = i < 2;
-            if is_a {
-                trace[row].new_a_carry_bits = F::from_u8(new_high_carry);
-                self.std.range_check(new_high_carry as i64, 1, self.a_range_id);
-            } else {
-                trace[row].new_e_carry_bits = F::from_u8(new_high_carry);
-                self.std.range_check(new_high_carry as i64, 1, self.e_range_id);
-            }
-
-            for j in 0..32 {
-                let bit = ((new_high >> j) & 1) as u8;
-                if is_a {
-                    trace[row].a[j] = F::from_u8(bit);
-                } else {
-                    trace[row].e[j] = F::from_u8(bit);
-                }
-            }
-            row -= 1;
-
-            if is_a {
-                trace[row].new_a_carry_bits = F::from_u8(new_low_carry);
-                self.std.range_check(new_low_carry as i64, 1, self.a_range_id);
-            } else {
-                trace[row].new_e_carry_bits = F::from_u8(new_low_carry);
-                self.std.range_check(new_low_carry as i64, 1, self.e_range_id);
-            }
-
-            for j in 0..32 {
-                let bit = ((new_low >> j) & 1) as u8;
-                if is_a {
-                    trace[row].a[j] = F::from_u8(bit);
-                } else {
-                    trace[row].e[j] = F::from_u8(bit);
-                }
-            }
-        }
-
-        // Perform the zero range checks
-        self.std.range_check(0, CLOCKS_LOAD_STATE as u64, self.a_range_id);
-        self.std.range_check(0, CLOCKS_LOAD_STATE as u64, self.e_range_id);
-
-        #[rustfmt::skip]
-        #[allow(clippy::too_many_arguments)]
-        fn compute_ae(old_a: u32, old_b: u32, old_c: u32, old_d: u32, old_e: u32, old_f: u32, old_g: u32, old_h: u32, w: u32, k: u32) -> (u64, u64) {
-            let s0 = rotate_right(old_a, 2) ^ rotate_right(old_a, 13) ^ rotate_right(old_a, 22);
-            let s1 = rotate_right(old_e, 6) ^ rotate_right(old_e, 11) ^ rotate_right(old_e, 25);
-            let t1 = (old_h as u64) + (s1 as u64) + (ch(old_e, old_f, old_g) as u64) + (k as u64) + (w as u64);
-            let t2 = (s0 as u64) + (maj(old_a, old_b, old_c) as u64);
-            let a = (t1 as u64) + (t2 as u64);
-            let e = (old_d as u64) + (t1 as u64);
-            (a, e)
-            // (s0 as u64, s1 as u64)
-        }
-
-        fn compute_w(old_w2: u32, old_w7: u32, old_w15: u32, old_w16: u32) -> u64 {
-            let s0 = rotate_right(old_w15, 7) ^ rotate_right(old_w15, 18) ^ shift_right(old_w15, 3);
-            let s1 = rotate_right(old_w2, 17) ^ rotate_right(old_w2, 19) ^ shift_right(old_w2, 10);
-            (s1 as u64) + (old_w7 as u64) + (s0 as u64) + (old_w16 as u64)
-        }
-
-        fn rotate_right(x: u32, n: u32) -> u32 {
-            (x >> n) | (x << (32 - n))
-        }
-
-        fn shift_right(x: u32, n: u32) -> u32 {
-            x >> n
-        }
-
-        fn maj(x: u32, y: u32, z: u32) -> u32 {
-            (x & y) ^ (x & z) ^ (y & z)
-        }
-
-        fn ch(x: u32, y: u32, z: u32) -> u32 {
-            (x & y) ^ (!x & z)
-        }
-    }
-
-    /// Computes the witness for a series of inputs and produces an `AirInstance`.
-    ///
-    /// # Arguments
-    /// * `sctx` - The setup context containing the setup data.
-    /// * `inputs` - A slice of operations to process.
-    ///
-    /// # Returns
-    /// An `AirInstance` containing the computed witness data.
-    pub fn compute_witness(&self, inputs: &[Vec<Sha256fInput>]) -> AirInstance<F> {
-        let num_rows = Sha256fTrace::<F>::NUM_ROWS;
-        let num_available_sha256fs = self.num_available_sha256fs;
-
-        // Check that we can fit all the sha256fs in the trace
-        let num_inputs = inputs.iter().map(|v| v.len()).sum::<usize>();
-        let num_rows_filled = num_inputs * CLOCKS;
-        let num_rows_needed = if num_inputs < num_available_sha256fs {
-            num_inputs * CLOCKS
-        } else if num_inputs == num_available_sha256fs {
-            num_rows
-        } else {
-            panic!(
-                "Exceeded available Sha256fs inputs: requested {}, but only {} are available.",
-                num_inputs, self.num_available_sha256fs
-            );
-        };
-
-        tracing::info!(
-            "··· Creating Sha256f instance [{} / {} rows filled {:.2}%]",
-            num_rows_needed,
-            num_rows,
-            num_rows_needed as f64 / num_rows as f64 * 100.0
-        );
-
-        timer_start_trace!(SHA256F_TRACE);
-        let mut sha256f_trace = Sha256fTrace::new_zeroes();
-
-        let mut trace_rows = sha256f_trace.buffer.as_mut_slice();
-        let mut par_traces = Vec::new();
-        let mut inputs_indexes = Vec::new();
-        for (i, inputs) in inputs.iter().enumerate() {
-            for (j, _) in inputs.iter().enumerate() {
-                let (head, tail) = trace_rows.split_at_mut(CLOCKS);
-                par_traces.push(head);
-                inputs_indexes.push((i, j));
-                trace_rows = tail;
-            }
-        }
-
-        // Fill the trace
-        par_traces.into_par_iter().enumerate().for_each(|(index, trace)| {
-            let input_index = inputs_indexes[index];
-            let input = &inputs[input_index.0][input_index.1];
-            self.process_input(input, trace);
-        });
-        timer_stop_and_log_trace!(SHA256F_TRACE);
-
-        timer_start_trace!(SHA256F_PADDING);
-        // Set a = e = w = 0 for the state and input rows
-        let zero_row = Sha256fTraceRow::<F>::default();
-
-        // precompute compute_ae() with initial a = e = 0 (PC_A and PC_E)
-        // compute_w() with w = 0 is equal to 0, nothing to do
-        let mut mid_rows = [Sha256fTraceRow::<F>::default(); 64];
-        for i in 0..64 {
-            let a = PC_A[i];
-            let e = PC_E[i];
-            let (a_carry, a) = ((a >> 32) as u8, (a & 0xFFFF_FFFF) as u32);
-            let (e_carry, e) = ((e >> 32) as u8, (e & 0xFFFF_FFFF) as u32);
-            mid_rows[i].new_a_carry_bits = F::from_u8(a_carry);
-            mid_rows[i].new_e_carry_bits = F::from_u8(e_carry);
-            for j in 0..32 {
-                let bit_a = ((a >> j) & 1) as u8;
-                let bit_e = ((e >> j) & 1) as u8;
-                mid_rows[i].a[j] = F::from_u8(bit_a);
-                mid_rows[i].e[j] = F::from_u8(bit_e);
-            }
-
-            self.std.range_check(
-                a_carry as i64,
-                (num_available_sha256fs - num_inputs) as u64,
-                self.a_range_id,
-            );
-            self.std.range_check(
-                e_carry as i64,
-                (num_available_sha256fs - num_inputs) as u64,
-                self.e_range_id,
-            );
-        }
-
-        // At the end, we should have that a === 4'and e === 4'e
-        let mut final_rows = [Sha256fTraceRow::<F>::default(); 4];
-        for i in 0..4 {
-            let a = (PC_A[60 + i] & 0xFFFF_FFFF) as u32;
-            let e = (PC_E[60 + i] & 0xFFFF_FFFF) as u32;
-            for j in 0..32 {
-                let bit_a = ((a >> j) & 1) as u8;
-                let bit_e = ((e >> j) & 1) as u8;
-                final_rows[i].a[j] = F::from_u8(bit_a);
-                final_rows[i].e[j] = F::from_u8(bit_e);
-            }
-        }
-
-        const CLOCKS_OP: usize = CLOCKS_LOAD_STATE + CLOCKS_LOAD_INPUT + CLOCKS_MIXING;
-        // The last (CLOCKS + NUM_NON_USABLE_ROWS) have CLK_0 desactivated, so
-        // a trace full of zeroes passes the constraints
-        for row in num_rows_filled..(num_rows - self.num_non_usable_rows - CLOCKS) {
-            let row_r = row % CLOCKS;
-            sha256f_trace[row] = if row_r < CLOCKS_LOAD_STATE {
-                zero_row
-            } else if row_r < CLOCKS_OP {
-                mid_rows[row_r - CLOCKS_LOAD_STATE]
-            } else {
-                final_rows[row_r - CLOCKS_OP]
-            };
-        }
-
-        // Perform the zero range checks
-        let count_zeros = (num_available_sha256fs - num_inputs)
-            * (CLOCKS_LOAD_STATE + CLOCKS_WRITE_STATE)
-            + CLOCKS
-            + self.num_non_usable_rows;
-        self.std.range_check(0, count_zeros as u64, self.a_range_id);
-        self.std.range_check(0, count_zeros as u64, self.e_range_id);
-        timer_stop_and_log_trace!(SHA256F_PADDING);
-
-        AirInstance::new_from_trace(FromTrace::new(&mut sha256f_trace))
-    }
-}
