+++ witness-computation/Cargo.toml
@@ -19,7 +19,7 @@ sm-mem = { path = "../state-machines/mem" }
 sm-rom = { path = "../state-machines/rom" }
 data-bus = { path = "../data-bus" }
 precomp-keccakf = { path = "../precompiles/keccakf" }
-precomp-sha256f = { path = "../precompiles/sha256f_direct" }
+precomp-sha256f = { path = "../precompiles/sha256f" }
 precomp-arith-eq = { path = "../precompiles/arith_eq" }
 zisk-pil = { path = "../pil" }
 ziskemu = { path = "../emulator" }
